test(htmltoslate): move whitespace tests
