+++ packages/html/src/lib/tests/htmlToSlate/configuration/elementTags.spec.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line @nx/enforce-module-boundaries
 import {
   htmlToSlate,
   HtmlToSlateConfig,
